requirements.txt
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg==3.0.14
psycopg2-binary==2.9.3
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
